feat: graceful Добавить принудительное завершение
